% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{chr2param}
\alias{chr2param}
\title{chr2param}
\usage{
chr2param(value)
}
\arguments{
\item{value}{value to transform}
}
\value{
transformed value
}
\description{
chr2param
}
\examples{
# some examples if you want to highlight the usage in the package
}
