% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{create_lookup_vector}
\alias{create_lookup_vector}
\title{create_lookup_vector}
\usage{
create_lookup_vector(lookup_table, key = "key", raw = "raw")
}
\arguments{
\item{lookup_table}{a lookup table}

\item{key}{name of key column}

\item{raw}{name of raw column}
}
\value{

}
\description{
create_lookup_vector
}
\examples{
# some examples if you want to highlight the usage in the package
}
