% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation_functions.R
\name{extract_plots_from_list}
\alias{extract_plots_from_list}
\title{title: plots_manipulation_functions.R
author: Vanessa Klaas
date: 2021-01-19}
\usage{
extract_plots_from_list(x)
}
\arguments{
\item{x}{A list object, potentially containing sublists.}
}
\value{
Returns a 'flat', single-level list of all the ggplot2 'plot-type' objects from within `x`, reaching recursively into sub-lists as needed. If there are no 'plot-type' objects, returns an empty list.
}
\description{
Takes a list (potentially containing sublists) and extracts all of the ggplot2 'plot-type' objects from that list into a simple list of 'plot-type' objects.
}
\details{
functions for manipulating plots
extract_plots_from_list
}
\note{
Whether an object is a ggplot2 'plot-type' object is defined here as an object with classes 'gg', 'gTree', or 'gtable'. https://gist.github.com/bschneidr/3f9862b8856f87d2b8f8a24878f76cc9
}
