% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_zhaw_theme.R
\docType{data}
\name{facets.default}
\alias{facets.default}
\title{facets.default, facet labeling size 12 and face bold}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 2.
}
\usage{
facets.default
}
\description{
facets.default, facet labeling size 12 and face bold
}
\keyword{datasets}
