% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\docType{data}
\name{fun_return}
\alias{fun_return}
\title{assign list elements that are returned by a function to variable names
usage: fun_return[result1, result2, result3] <-
          function_with_list_of_3_return_values(...)}
\format{
An object of class \code{result} of length 1.
}
\usage{
fun_return
}
\value{
assigns a list of return values to single variables
}
\description{
assign list elements that are returned by a function to variable names
usage: fun_return[result1, result2, result3] <-
          function_with_list_of_3_return_values(...)
}
\keyword{datasets}
