% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{\%+\%}
\alias{\%+\%}
\title{inline operator for string concatenation based on paste0}
\usage{
x \%+\% y
}
\arguments{
\item{x}{string1}

\item{y}{string2}
}
\value{
paste0(x, y)
}
\description{
inline operator for string concatenation based on paste0
}
