% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation_functions.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid_arrange_shared_legend}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\arguments{
\item{...}{plots to put together}

\item{ncol}{number of columns}

\item{nrow}{number or rows}

\item{position}{position of legend}
}
\value{
plot with subplots and shared legend
}
\description{
grid_arrange_shared_legend
}
