% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_zhaw_theme.R
\docType{data}
\name{kMinor.blank}
\alias{kMinor.blank}
\title{kMinor.blank x and y axis minor blank}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
kMinor.blank
}
\description{
kMinor.blank x and y axis minor blank
}
\keyword{datasets}
