% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_zhaw_theme.R
\docType{data}
\name{kXLab.long}
\alias{kXLab.long}
\alias{kX.empty}
\title{kXLab.long x-axis with long labels, 45 degree}
\format{
An object of class \code{element_text} (inherits from \code{element}) of length 11.

An object of class \code{theme} (inherits from \code{gg}) of length 3.
}
\usage{
kXLab.long

kX.empty
}
\description{
usage: theme(axis.text.x = kXLab.long) +
}
\keyword{datasets}
