% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_io.R
\name{load_lookup_table}
\alias{load_lookup_table}
\title{load_lookup_table}
\usage{
load_lookup_table(file_path)
}
\arguments{
\item{file_path}{path to file including file name}
}
\value{
tibble
}
\description{
load_lookup_table
}
