% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{normalize_blanks_tolower}
\alias{normalize_blanks_tolower}
\title{normalize_blanks_tolower}
\usage{
normalize_blanks_tolower(x)
}
\arguments{
\item{x}{string or vector of strings}
}
\value{
normalized string or vector of normalized strings
}
\description{
normalize_blanks_tolower
}
\examples{
# some examples if you want to highlight the usage in the package
}
