% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{normalize_text}
\alias{normalize_text}
\title{normalize_text}
\usage{
normalize_text(x, lookup_table, pattern = "raw", replacement = "key")
}
\arguments{
\item{x}{a vector of values}

\item{lookup_table}{a lookup table}

\item{pattern}{name of raw/pattern column}

\item{replacement}{name of key/replacement column}
}
\value{
vector of normalized text values
}
\description{
normalize_text
}
\examples{
# some examples if you want to highlight the usage in the package
}
