% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{\%>\%}
\alias{\%>\%}
\title{title: misc_functions.R
author: Vanessa Klaas
date: 2021-01-18}
\arguments{
\item{lhs, rhs}{specify what lhs and rhs are}
}
\description{
little helper functions
Pipe
}
\details{
Put description here
}
\examples{
# some examples if you want to highlight the usage in the package
}
