% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_management_functions.R
\name{PlotFileName}
\alias{PlotFileName}
\title{title: plots_management_functions.R
author: Vanessa Klaas
date: 2021-01-19}
\usage{
PlotFileName(
  prefix,
  middle,
  postfix,
  sep = "_",
  rootdir = "plots/",
  suffix = "png"
)
}
\arguments{
\item{prefix}{a prefix}

\item{middle}{middle}

\item{postfix}{postfix}

\item{sep}{separator}

\item{rootdir}{root directory for plots}

\item{suffix}{file extension}
}
\value{
filename with complete path
}
\description{
functions for plots
PlotFileName
}
