% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_labeling.R
\name{relabel_values}
\alias{relabel_values}
\title{title: functions_labeling.R
author: Vanessa Klaas
date: 2021-01-18}
\usage{
relabel_values(x, lookup_table, pattern = "key", replacement = "label")
}
\arguments{
\item{x}{vector of strings to relabel}

\item{lookup_table}{df containing keys and labels}

\item{pattern}{name of key column}

\item{replacement}{name of label column}
}
\value{
vector with relabeled values
}
\description{
functions for labeling for plots
relabel_values
}
\examples{
# some examples if you want to highlight the usage in the package
}
