% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{rename_cols}
\alias{rename_cols}
\title{rename_cols}
\usage{
rename_cols(lookup_table, key = "key", raw = "raw")
}
\arguments{
\item{lookup_table}{a lookup table}

\item{key}{name of key column}

\item{raw}{name of raw column}
}
\value{
lookup table with renamed columns
}
\description{
rename_cols
}
\examples{
# some examples if you want to highlight the usage in the package
}
