% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_management_functions.R
\name{SavePlot}
\alias{SavePlot}
\title{save Plot}
\usage{
SavePlot(
  g,
  size,
  prefix,
  middle,
  postfix,
  sep = "_",
  rootdir = "plots/",
  suffix = "png",
  dpi = 300,
  width = NULL,
  height = NULL,
  units = "cm"
)
}
\arguments{
\item{g}{the ggplot object to save}

\item{size}{the plot size, possible options are: a4', 'a4.hor', 'a4.hor.half', 'col.1' and more}

\item{prefix}{a prefix}

\item{middle}{middle}

\item{postfix}{postfix}

\item{sep}{separator}

\item{rootdir}{root directory for plots}

\item{suffix}{file extension}

\item{dpi}{resolution, default=300}

\item{width}{custom width}

\item{height}{custom height}

\item{units}{custom unit}
}
\value{
None
}
\description{
save Plot
}
