% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_zhaw_theme.R
\name{theme_zhaw}
\alias{theme_zhaw}
\title{theme_zhaw}
\usage{
theme_zhaw(
  base_size = 9,
  base_family = "Lato",
  panel.grid.major.x.blank = FALSE,
  panel.grid.minor.y.blank = FALSE
)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{panel.grid.major.x.blank}{if TRUE, no major x gridlines}

\item{panel.grid.minor.y.blank}{if TRUE, no minor y gridlines}
}
\value{
zhaw theme
}
\description{
theme_zhaw
}
