% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{create_lookup_vector}
\alias{create_lookup_vector}
\title{create_lookup_vector}
\usage{
create_lookup_vector(lookup_table, pattern = "raw", replacement = "key")
}
\arguments{
\item{lookup_table}{a lookup table}

\item{pattern}{name of raw/pattern column}

\item{replacement}{name of key/replacement column}
}
\value{
named vector to use with str_replace_all
}
\description{
create_lookup_vector
}
\examples{
# some examples if you want to highlight the usage in the package
}
