% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{normalize_text}
\alias{normalize_text}
\title{normalize_text}
\usage{
normalize_text(x, lookup_table, key = "key", raw = "raw")
}
\arguments{
\item{x}{a vector of values}

\item{lookup_table}{a lookup table}

\item{key}{name of key column}

\item{raw}{name of raw column}
}
\value{

}
\description{
normalize_text
}
\examples{
# some examples if you want to highlight the usage in the package
}
