% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_management_functions.R
\name{PlotFileName}
\alias{PlotFileName}
\title{title: plots_management_functions.R
author: Vanessa Klaas
date: 2021-01-19}
\usage{
PlotFileName(prefix, camid, rootdir = "plots/", suffix = "png", postfix = "")
}
\arguments{
\item{prefix}{a prefix}

\item{camid}{campaign id}

\item{rootdir}{root directory for plots}

\item{suffix}{file extension}

\item{postfix}{postfix for filename}
}
\value{
filename with complete path
}
\description{
functions for plots
PlotFileName
}
