% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{rename_cols}
\alias{rename_cols}
\title{title: functions_cleaning.R
author: Vanessa Klaas
date: 2021-01-13}
\usage{
rename_cols(lookup_table, pattern = "key", replacement = "raw")
}
\arguments{
\item{lookup_table}{a lookup table}

\item{key}{name of key column}

\item{raw}{name of raw column}
}
\value{
lookup table with renamed columns
}
\description{
clean, normalize, and unify strings
rename_cols
}
\examples{
# some examples if you want to highlight the usage in the package
}
