% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_management_functions.R
\name{SavePlot}
\alias{SavePlot}
\title{save Plot}
\usage{
SavePlot(
  prefix,
  camid,
  g,
  size,
  rootdir = "plots/",
  suffix = "png",
  postfix = "",
  dpi = 300
)
}
\arguments{
\item{prefix}{prefix of filename}

\item{camid}{campaign ID}

\item{g}{the ggplot object to save}

\item{size}{the plot size, possible options are: a4', 'a4.hor', 'a4.hor.half', 'col.1' and more}

\item{rootdir}{the directory in which the file shall be saved, default plots/}

\item{suffix}{file extensions / format}

\item{postfix}{string inserted after the camid, default png}

\item{dpi}{resolution, default=300}
}
\value{
None
}
\description{
save Plot
}
